+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/TableKey.java
@@ -1,0 +1,6 @@
+package org.openprovenance.prov.service;
+
+public class TableKey {
+    public String isA;
+    public Integer ID;
+}