+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/TableKeyList.java
@@ -1,0 +1,9 @@
+package org.openprovenance.prov.service;
+
+import java.util.LinkedList;
+import java.util.List;
+
+public class TableKeyList {
+    public List<TableKey> key=new LinkedList<>();
+
+}