+++ RELEASES.md
@@ -8,6 +8,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates: submission of csv with json results and reconstructed csv_log
 * service-templates: visualising prov for a template instantiation
 * service-templates: plead workflow, run locally and remotely
+* service-templates: extracting provenance of a workflow run
 
 `2.0.5`
 * prov-template-library: cleaned up pom.xml